Refactor into smaller modules
